+++ packages/ui/src/components/sign-in/sign-in.tsx
@@ -4,6 +4,7 @@ import * as SignIn from '@clerk/elements/sign-in';
 
 import { EmailField } from '~/common/EmailField';
 import { PasswordField } from '~/common/PasswordField';
+import { PhoneNumberField } from '~/common/PhoneNumberField';
 import { PROVIDERS } from '~/constants';
 import { getEnabledSocialConnectionsFromEnvironment } from '~/hooks/getEnabledSocialConnectionsFromEnvironment';
 import { Button } from '~/primitives/button';
@@ -60,6 +61,8 @@ export function SignInComponent() {
                     </Connection.Root>
                     <Seperator>or</Seperator>
                     <div className='space-y-4'>
+                      {/* @ts-ignore Expected https://github.com/clerk/javascript/blob/12f78491d6b10f2be63891f8a7f76fc6acf37c00/packages/clerk-js/src/ui/elements/PhoneInput/PhoneInput.tsx#L248-L249 */}
+                      <PhoneNumberField locationBasedCountryIso={clerk.__internal_country} />
                       <EmailField disabled={isGlobalLoading} />
                       <PasswordField disabled={isGlobalLoading} />
                     </div>
